Renamed ns GetTagActionDetails to GetActionDetails and GetTagActions to GetActions
