+++ src/Equinor.Procosys.Preservation.Query/GetActionDetails/GetActionDetailsQueryHandler.cs
@@ -10,7 +10,7 @@ using Microsoft.EntityFrameworkCore;
 using ServiceResult;
 using Action = Equinor.Procosys.Preservation.Domain.AggregateModels.ProjectAggregate.Action;
 
-namespace Equinor.Procosys.Preservation.Query.GetTagActionDetails
+namespace Equinor.Procosys.Preservation.Query.GetActionDetails
 {
     public class GetActionDetailsQueryHandler : IRequestHandler<GetActionDetailsQuery, Result<ActionDetailsDto>>
     {
@@ -23,12 +23,12 @@ namespace Equinor.Procosys.Preservation.Query.GetTagActionDetails
             var dto = await
                 (from a in _context.QuerySet<Action>()
                      // also join tag to return null if request.TagId not exists
-                 join tag in _context.QuerySet<Tag>() on request.TagId equals tag.Id
+                     join tag in _context.QuerySet<Tag>() on EF.Property<int>(a, "TagId") equals tag.Id
                  join createdUser in _context.QuerySet<Person>()
                      on EF.Property<int>(a, "CreatedById") equals createdUser.Id
                  from closedUser in _context.QuerySet<Person>()
                      .Where(p => p.Id == EF.Property<int>(a, "ClosedById")).DefaultIfEmpty() // left join
-                 where a.Id == request.ActionId
+                 where tag.Id == request.TagId && a.Id == request.ActionId
                  select new Dto
                  {
                      Action = a,
