+++ src/tests/Equinor.Procosys.Preservation.Query.Tests/GetActionDetails/GetActionDetailsQueryTests.cs
@@ -1,4 +1,4 @@
-﻿using Equinor.Procosys.Preservation.Query.GetTagActionDetails;
+﻿using Equinor.Procosys.Preservation.Query.GetActionDetails;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Equinor.Procosys.Preservation.Query.Tests.GetActionDetails
