+++ src/tests/Equinor.Procosys.Preservation.Query.Tests/GetActions/GetActionsQueryTests.cs
@@ -1,15 +1,15 @@
-﻿using Equinor.Procosys.Preservation.Query.GetTagActions;
+﻿using Equinor.Procosys.Preservation.Query.GetActions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Equinor.Procosys.Preservation.Query.Tests.GetTagActions
+namespace Equinor.Procosys.Preservation.Query.Tests.GetActions
 {
     [TestClass]
-    public class GetTagActionsQueryTests
+    public class GetActionsQueryTests
     {
         [TestMethod]
         public void Constructor_SetsProperties()
         {
-            var dut = new GetTagActionsQuery(1337);
+            var dut = new GetActionsQuery(1337);
 
             Assert.AreEqual(1337, dut.TagId);
         }
